requirements.txt
langchain
langchain-groq
python-dotenv
gunicorn